LayoutTests/http/tests/security/xss-DENIED-invalid-domain-change-expected.txt

CONSOLE MESSAGE: line 17: SecurityError: Attempted to use a non-registrable domain.

PASS: cross-site not access allowed
